requirements.txt
Flask==3.1.2
flask-cors==6.0.1
numpy==2.3.2
pillow==11.3.0
torch==2.8.0
transformers==4.56.0
colorlog==6.9.0